Fix mergesort and keep comments (for now)
Made it so the "secondIndex" in mergeRanges doesn't include the midpoint
(by adding 1 to it) and made it so the firstIndex was inclusive with the
midpoint (made it so the conditionals were "firstIndex <= midPoint").
This, at the very least, helps the base case when there is only one
integer in both sections being merged. This is because the midpoint is
equal to 0 in this case (it could be the case that (startIndex +
endIndex is wrong, but I'm bot sure and this works for now).