Add print statements and for loop to add elements to "destination"
The print statements check to see what is inside all of the inputs. Then
destination is going to eventually be the final result. So, there is a
for loop that puts everything into the destination array and then also
at the end there is another print statement that checks the contents of
destination.